make ChoiceType be compatible with Enum type.
