+++ docs/data_types.rst
@@ -45,13 +45,6 @@ EncryptedType
 
 .. autoclass:: EncryptedType
 
-EnumType
-^^^^^^^^
-
-.. module:: sqlalchemy_utils.types.enum
-
-.. autoclass:: EnumType
-
 JSONType
 ^^^^^^^^
