+++ sqlalchemy_utils/__init__.py
@@ -65,7 +65,6 @@ from .types import (
     DateTimeRangeType,
     EmailType,
     EncryptedType,
-    EnumType,
     instrumented_list,
     InstrumentedList,
     IntRangeType,
@@ -145,7 +144,6 @@ __all__ = (
     DateTimeRangeType,
     EmailType,
     EncryptedType,
-    EnumType,
     ExpressionParser,
     ImproperlyConfigured,
     InstrumentedList,
