+++ tests/types/test_choice.py
@@ -1,7 +1,8 @@
 from flexmock import flexmock
-from pytest import raises
+from pytest import raises, mark
 import sqlalchemy as sa
 from sqlalchemy_utils import ChoiceType, Choice, ImproperlyConfigured
+from sqlalchemy_utils.types.choice import Enum
 from tests import TestCase
 
 
@@ -76,3 +77,53 @@ class TestChoiceTypeWithCustomUnderlyingType(TestCase):
     def test_init_type(self):
         type_ = ChoiceType([(1, u'something')], impl=sa.Integer)
         assert type_.impl == sa.Integer
+
+
+@mark.skipif('Enum is None')
+class TestEnumType(TestCase):
+    def create_models(self):
+        class OrderStatus(Enum):
+            unpaid = 1
+            paid = 2
+
+        class Order(self.Base):
+            __tablename__ = 'order'
+            id_ = sa.Column(sa.Integer, primary_key=True)
+            status = sa.Column(
+                ChoiceType(OrderStatus, impl=sa.Integer()),
+                default=OrderStatus.unpaid)
+
+            def __repr__(self):
+                return 'Order(%r, %r)' % (self.id_, self.status)
+
+            def pay(self):
+                self.status = OrderStatus.paid
+
+        self.OrderStatus = OrderStatus
+        self.Order = Order
+
+    def test_parameter_processing(self):
+        order = self.Order()
+
+        self.session.add(order)
+        self.session.commit()
+
+        order = self.session.query(self.Order).first()
+        assert order.status is self.OrderStatus.unpaid
+        assert order.status.value == 1
+
+        order.pay()
+        self.session.commit()
+
+        order = self.session.query(self.Order).first()
+        assert order.status is self.OrderStatus.paid
+        assert order.status.value == 2
+
+    def test_parameter_coercing(self):
+        order = self.Order()
+        order.status = 2
+
+        self.session.add(order)
+        self.session.commit()
+
+        assert order.status is self.OrderStatus.paid
